\name{linnerud}
\alias{linnerud}
\title{Multivariate dataset for PCR}
\usage{data(linnerud)}
\references{J.E. Jackson, "A user's guide to principal components, p.267 
(1991).}
\format{A dataframe with an X matrix of 20 objects and 3 variables, and
a Y matrix with 3 responses.}
\description{This dataset is used by Jackson to illustrate PCR, ridge
regression and other multivariate regression methods.}
\keyword{datasets}
