\name{NIR}
\alias{NIR}
\title{Multivariate dataset of NIR spectra}
\usage{data(NIR)}
\references{H. Swierenga et al., Chemom. Intell. Lab. Syst. 49 (1999) 1-17.}
\format{A dataframe with matrices Xtrain (21x268) and Xtest (7x268), and 
two vectors Ytrain and Ytest.}
\description{A training set consisting of 21 NIR spectra of
PET yarns, measured at 268 wavelengths, and 21 corresponding
densities. A training set of 7 samples is also provided.
Many thanks to Erik Swierenga...}
\keyword{datasets}
