\name{sensory}
\alias{sensory}
\title{Multivariate dataset of sensory data}
\usage{data(sensory)}
\references{B.G.M. Vandeginste et al., Handbook of Chemometrics and
  Qualimetrics: Part B, Elsevier (1998), Tables 35.1 and 35.4.}
\format{A dataframe with matrices Quality (16 x 5) and Panel (16 x 6).}
\description{Panel gives scores from a sensory panel on a set of 6
  attributes for 16 olive oil samples. Quality gives the results of 5
  physico-chemical quality parameters. The first five oils are Greek,
  the next five are Italian and the last six are Spanish. This data set
  illustrates the qualitative capabilities of PLS.}
\keyword{datasets}
